Project/requirements.txt
asgiref==3.3.1
azure-common==1.1.26
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.1.5
django-crispy-forms==1.10.0
django-storages==1.11.1
idna==2.10
Pillow==8.1.0
pip==20.2.3
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
setuptools==49.2.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
